requirements.txt
flask
flask_socketio
uuid
python-dotenv